requirements.txt
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
python-dotenv==1.0.1
yt-dlp==2025.1.15
openai-whisper @ git+https://github.com/openai/whisper.git
google-genai